User controllers error handling

+++ app/Http/Controllers/UserController.php
@@ -43,26 +43,32 @@ class UserController extends Controller
             if (!$request->password) array_push($response['error_msg'], 'Password is required');
             if (!$request->user_name) array_push($response['error_msg'], 'User name is required');
             if (!count($response['error_msg']) > 0) {
-                //TRY CATCH?? 
-                $user = User::where('email', '=', $request->email);
+                //TODO - TO TEST 
+                try {
+                    $user = User::where('email', '=', $request->email);
 
-                if (!$user->count()) {
-                    try {
-                        $user = new User();
-                        $user->email = $request->email;
-                        $user->password = hash('sha256', $request->password);
-                        $user->user_name = $request->user_name;
-                        $token = uniqid() . $user->email;
-                        $user->token = hash('sha256', $token);
-                        $user->save();
-                        $response = array('code' => 200, 'user' => $user, 'msg' => 'User created');
-                    } catch (\Exception $exception) {
-                        $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                    if (!$user->count()) {
+                        try {
+                            $user = new User();
+                            $user->email = $request->email;
+                            $user->password = hash('sha256', $request->password);
+                            $user->user_name = $request->user_name;
+                            $token = uniqid() . $user->email;
+                            $user->token = hash('sha256', $token);
+                            $user->save();
+                            $response = array('code' => 200, 'user' => $user, 'msg' => 'User created');
+                        } catch (\Exception $exception) {
+                            $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                        }
+
+                    } else {
+                        $response = array('code' => 400, 'error_msg' => "Email already registered");
                     }
 
-                } else {
-                    $response = array('code' => 400, 'error_msg' => "Email already registered");
+                } catch (\Throwable $th) {
+                    $response = array('code' => 500, 'error_msg' => $exception->getMessage());
                 }
+
             }
 
         } else {
@@ -76,19 +82,26 @@ class UserController extends Controller
     public function updateUser(Request $request, $id)
     {
         $response = array('code' => 400, 'error_msg' => []);
-        //TODO - TRY CATCH??
-        $user = User::find($id);
-
-        if (isset($request) && isset($id) && !empty($user)) {
+        
+        if (isset($request) && isset($id)){
+            //TODO - TO TEST
             try {
-                $user->email = $request->email ? $request->email : $user->email;
-                $user->password = $request->password ? hash('sha256', $request->password) : $user->password;
-                $user->user_name = $request->user_name ? $request->user_name : $user->user_name;
-                $user->location = $request->location ? $request->location : $user->location;
-                $user->picture = $request->picture ? $request->picture : $user->picture;
-                $user->save();
-                $response = array('code' => 200, 'msg' => 'User updated');
-            } catch (\Exception $exception) {
+                $user = User::find($id);
+
+                if (!empty($user)) {
+                    try {
+                        $user->email = $request->email ? $request->email : $user->email;
+                        $user->password = $request->password ? hash('sha256', $request->password) : $user->password;
+                        $user->user_name = $request->user_name ? $request->user_name : $user->user_name;
+                        $user->location = $request->location ? $request->location : $user->location;
+                        $user->picture = $request->picture ? $request->picture : $user->picture;
+                        $user->save();
+                        $response = array('code' => 200, 'msg' => 'User updated');
+                    } catch (\Exception $exception) {
+                        $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                    }
+                }
+            } catch (\Throwable $th) {
                 $response = array('code' => 500, 'error_msg' => $exception->getMessage());
             }
 
@@ -125,21 +138,28 @@ class UserController extends Controller
     //Delete an specific user by ID
     public function deleteUser(Request $request, $id)
     {
-        //TODO - TRY CATCH??
-        $user = User::find($id);
-
-        if (!empty($user)  && $request->user('api')->admin_user === 1) {
+        if (isset($request) && isset($id)) {
+            //TODO - TO TEST
             try {
-                $user->delete();
-                $response = array('code' => 200, 'msg' => 'User deleted');
-            } catch (\Exception $exception) {
+                $user = User::find($id);
+
+                if (!empty($user)  && $request->user('api')->admin_user === 1) {
+                    try {
+                        $user->delete();
+                        $response = array('code' => 200, 'msg' => 'User deleted');
+                    } catch (\Exception $exception) {
+                        $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                    }
+        
+                } else {
+                    $response = array('code' => 401, 'error_msg' => 'Unautorized');
+                }
+
+            } catch (\Throwable $th) {
                 $response = array('code' => 500, 'error_msg' => $exception->getMessage());
             }
-
-        } else {
-            $response = array('code' => 401, 'error_msg' => 'Unautorized');
         }
-
+       
        return response($response,$response['code']);
     }
 
@@ -149,25 +169,30 @@ class UserController extends Controller
         $response = array('code' => 400, 'error_msg' => []);
 
         if ($request->email && $request->password) {
-            //TODO - TRY CATCH??
-            $user = User::where('email', "$request->email")->first();
+            //TODO - TO TEST
+            try {
+                $user = User::where('email', "$request->email")->first();
 
-            if (!empty($user)) {
-                if ($user->password === hash('sha256', $request->password)) {
-                    try {
-                        $token = uniqid() . $user->email;
-                        $user->token = hash('sha256', $token);
-                        $user->save();
-                        $response = array('code' => 200, 'user' => $user, 'msg' => 'Login successful',);
-                    } catch (\Exception $exception) {
-                        $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                if (!empty($user)) {
+                    if ($user->password === hash('sha256', $request->password)) {
+                        try {
+                            $token = uniqid() . $user->email;
+                            $user->token = hash('sha256', $token);
+                            $user->save();
+                            $response = array('code' => 200, 'user' => $user, 'msg' => 'Login successful',);
+                        } catch (\Exception $exception) {
+                            $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                        }
+                    } else {
+                        $response['error_msg'] = 'Wrong password';
                     }
                 } else {
-                    $response['error_msg'] = 'Wrong password';
+                    $response['error_msg'] = 'User not found';
                 }
-            } else {
-                $response['error_msg'] = 'User not found';
+            } catch (\Throwable $th) {
+                $response = array('code' => 500, 'error_msg' => $exception->getMessage());
             }
+            
 
         } else {
             $response['error_msg'] = 'Email and password are required';
@@ -181,50 +206,59 @@ class UserController extends Controller
     {
         $response = array('code' => 400, 'error_msg' => []);
 
-        try {
-            //User object
-            $user = User::where('email', '=', $request->email)->first();
-        } catch (\Throwable $th) {
-            $response = array('code' => 400, 'error_msg' => 'User not found');
-        }
-
-        //Checking if the email exist
-        if (!empty($user)) {
-            //New password of the user
-            $newPass = $this->rand_string(8);
-            //User data that will be used on the email
-            $email = $user->email;
-            $name = $user->user_name;
-
-            //Hash the new password  
-            $password = hash('sha256', $newPass);
-
-            //Save the new password to the user
-            $user->password = $password;
-            $user->save();
-
-            //Email sender and relative data 
-            $data = [
-                'name' => $name,
-                'password' => $newPass,
-            ];
-
-            $subject = "PetIt App - Reset password request";
-            $from =  env("MAIL_USERNAME");
+        if (isset($request)){
 
             try {
-                //Send Mail
-                $mailMsg = Mail::send('mail', ["data" => $data], function ($msg) use ($subject, $email, $from) {
-                    $msg->from($from, "🐾 PetIt App 🐾");
-                    $msg->subject($subject);
-                    $msg->to($email);
-                });
-                $response = array('code' => 200, 'error_msg' => 'Email sended!');
+                //User object
+                $user = User::where('email', '=', $request->email)->first();
+
+                //Checking if the email exist
+                if (!empty($user)) {
+                    //New password of the user
+                    $newPass = $this->rand_string(8);
+                    //User data that will be used on the email
+                    $email = $user->email;
+                    $name = $user->user_name;
+        
+                    //Hash the new password  
+                    $password = hash('sha256', $newPass);
+        
+                    //Save the new password to the user
+                    $user->password = $password;
+                    $user->save();
+        
+                    //Email sender and relative data 
+                    $data = [
+                        'name' => $name,
+                        'password' => $newPass,
+                    ];
+        
+                    $subject = "PetIt App - Reset password request";
+                    $from =  env("MAIL_USERNAME");
+        
+                    try {
+                        //Send Mail
+                        $mailMsg = Mail::send('mail', ["data" => $data], function ($msg) use ($subject, $email, $from) {
+                            $msg->from($from, "🐾 PetIt App 🐾");
+                            $msg->subject($subject);
+                            $msg->to($email);
+                        });
+                        $response = array('code' => 200, 'error_msg' => 'Email sended!');
+                    } catch (\Throwable $th) {
+                        // $response = array('code' => 400, 'error_msg' => 'Error sending the message...');
+                        $response = array('code' => 500, 'error_msg' => $exception->getMessage());
+                    }
+
+                } else {
+                    $response = array('code' => 400, 'error_msg' => 'User not found');
+                }
+
             } catch (\Throwable $th) {
-                $response = array('code' => 400, 'error_msg' => 'Error sending the message...');
+                $response = array('code' => 500, 'error_msg' => $exception->getMessage());
             }
+    
         } else {
-            $response = array('code' => 400, 'error_msg' => 'User not found');
+            $response = array('code' => 400, 'error_msg' => 'No email received');
         }
 
        return response($response,$response['code']);
